Generic tracing; removed iso14a_XX-functions, removed traceLen as global varible
