+++ armsrc/BigBuf.h
@@ -29,6 +29,9 @@ extern uint8_t *BigBuf_malloc(uint16_t);
 extern void BigBuf_free(void);
 extern void BigBuf_free_keep_EM(void);
 
-extern uint16_t traceLen;
+uint16_t BigBuf_get_traceLen(void);
+void clear_trace();
+void set_tracing(bool enable);
+bool RAMFUNC LogTrace(const uint8_t *btBytes, uint16_t iLen, uint32_t timestamp_start, uint32_t timestamp_end, uint8_t *parity, bool readerToTag);
 
 #endif /* __BIGBUF_H */
