+++ armsrc/mifarecmd.c
@@ -41,7 +41,7 @@ void MifareReadBlock(uint8_t arg0, uint8_t arg1, uint8_t arg2, uint8_t *datain)
 	pcs = &mpcs;
 
 	// clear trace
- 	iso14a_clear_trace();
+	clear_trace();
 	iso14443a_setup(FPGA_HF_ISO14443A_READER_LISTEN);
 
 	LED_A_ON();
@@ -98,7 +98,7 @@ void MifareUC_Auth1(uint8_t arg0, uint8_t *datain){
 	LED_B_OFF();
 	LED_C_OFF();
     
-	iso14a_clear_trace();
+	clear_trace();
 	iso14443a_setup(FPGA_HF_ISO14443A_READER_LISTEN);
 
 	if(!iso14443a_select_card(uid, NULL, &cuid)) {
@@ -162,7 +162,7 @@ void MifareUReadBlock(uint8_t arg0,uint8_t *datain)
 	LED_B_OFF();
 	LED_C_OFF();
     
-	iso14a_clear_trace();
+	clear_trace();
 	iso14443a_setup(FPGA_HF_ISO14443A_READER_LISTEN);
     
 	int len = iso14443a_select_card(uid, NULL, &cuid);
@@ -213,7 +213,7 @@ void MifareReadSector(uint8_t arg0, uint8_t arg1, uint8_t arg2, uint8_t *datain)
 	pcs = &mpcs;
 
 	// clear trace
- 	iso14a_clear_trace();
+	clear_trace();
 
 	iso14443a_setup(FPGA_HF_ISO14443A_READER_LISTEN);
 
@@ -276,7 +276,7 @@ void MifareUReadCard(uint8_t arg0, int arg1, uint8_t *datain)
 	if (MF_DBGLEVEL >= MF_DBG_ALL) 
 		Dbprintf("Pages %d",Pages);
 	
-	iso14a_clear_trace();
+	clear_trace();
 	iso14443a_setup(FPGA_HF_ISO14443A_READER_LISTEN);
 
 	int len = iso14443a_select_card(uid, NULL, &cuid);
@@ -350,7 +350,7 @@ void MifareWriteBlock(uint8_t arg0, uint8_t arg1, uint8_t arg2, uint8_t *datain)
 	pcs = &mpcs;
 
 	// clear trace
-	iso14a_clear_trace();
+	clear_trace();
 
 	iso14443a_setup(FPGA_HF_ISO14443A_READER_LISTEN);
 
@@ -411,7 +411,7 @@ void MifareUWriteBlock(uint8_t arg0, uint8_t *datain)
 	uint8_t uid[10] = {0x00};
         uint32_t cuid;
 
-        iso14a_clear_trace();
+		clear_trace();
 		iso14443a_setup(FPGA_HF_ISO14443A_READER_LISTEN);
 
         LED_A_ON();
@@ -458,7 +458,7 @@ void MifareUWriteBlock_Special(uint8_t arg0, uint8_t *datain)
 	uint8_t uid[10] = {0x00};
 	uint32_t cuid;
 
-	iso14a_clear_trace();
+	clear_trace();
 	iso14443a_setup(FPGA_HF_ISO14443A_READER_LISTEN);
 
 	LED_A_ON();
@@ -537,8 +537,8 @@ void MifareNested(uint32_t arg0, uint32_t arg1, uint32_t calibrate, uint8_t *dat
 	// free eventually allocated BigBuf memory
 	BigBuf_free();
 	// clear trace
-	iso14a_clear_trace();
-	iso14a_set_tracing(false);
+	clear_trace();
+	set_tracing(false);
 	
 	iso14443a_setup(FPGA_HF_ISO14443A_READER_LISTEN);
 
@@ -709,7 +709,7 @@ void MifareNested(uint32_t arg0, uint32_t arg1, uint32_t calibrate, uint8_t *dat
 
 	FpgaWriteConfWord(FPGA_MAJOR_MODE_OFF);
 	LEDsoff();
-	iso14a_set_tracing(TRUE);
+	set_tracing(TRUE);
 }
 
 //-----------------------------------------------------------------------------
@@ -738,8 +738,8 @@ void MifareChkKeys(uint8_t arg0, uint8_t arg1, uint8_t arg2, uint8_t *datain)
 	MF_DBGLEVEL = MF_DBG_NONE;
 	
 	// clear trace
-	iso14a_clear_trace();
-	iso14a_set_tracing(TRUE);
+	clear_trace();
+	set_tracing(TRUE);
 
 	iso14443a_setup(FPGA_HF_ISO14443A_READER_LISTEN);
 
@@ -829,8 +829,8 @@ void MifareECardLoad(uint32_t arg0, uint32_t arg1, uint32_t arg2, uint8_t *datai
 	uint8_t uid[10];
 
 	// clear trace
-	iso14a_clear_trace();
-	iso14a_set_tracing(false);
+	clear_trace();
+	set_tracing(false);
 	
 	iso14443a_setup(FPGA_HF_ISO14443A_READER_LISTEN);
 
@@ -931,8 +931,8 @@ void MifareCSetBlock(uint32_t arg0, uint32_t arg1, uint32_t arg2, uint8_t *datai
 		LED_B_OFF();
 		LED_C_OFF();
 	
-		iso14a_clear_trace();
-		iso14a_set_tracing(TRUE);
+		clear_trace();
+		set_tracing(TRUE);
 		iso14443a_setup(FPGA_HF_ISO14443A_READER_LISTEN);
 	}
 
@@ -1049,8 +1049,8 @@ void MifareCGetBlock(uint32_t arg0, uint32_t arg1, uint32_t arg2, uint8_t *datai
 		LED_B_OFF();
 		LED_C_OFF();
 	
-		iso14a_clear_trace();
-		iso14a_set_tracing(TRUE);
+		clear_trace();
+		set_tracing(TRUE);
 		iso14443a_setup(FPGA_HF_ISO14443A_READER_LISTEN);
 	}
 
@@ -1136,7 +1136,7 @@ void Mifare_DES_Auth1(uint8_t arg0, uint8_t *datain){
 	uint8_t uid[10] = {0x00};
 	uint32_t cuid;
     
-	iso14a_clear_trace();
+	clear_trace();
 	iso14443a_setup(FPGA_HF_ISO14443A_READER_LISTEN);
 
 	int len = iso14443a_select_card(uid, NULL, &cuid);
